Add API GamesController and update routes
